+++ CHANGELOG.md
@@ -7,6 +7,9 @@ All notable changes to this project will be documented in this file. This change
 - Add -keep-gcl option to the clean script
 - Ignore `*.swp` files
 
+### Changed
+- Re-enable FTL JIT on macOS ([#820](https://github.com/planck-repl/planck/issues/820))
+
 ### Fixed
 - Switch `strncpy` to `memcpy` to avoid GCC warning
 - Backslash return return should produce "\n" ([661](https://github.com/planck-repl/planck/issues/661))